# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/category": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting all category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get All Category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipper",
                        "name": "Shipper",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "parent_id",
                        "name": "parent_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "all",
                        "name": "all",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "with_products",
                        "name": "with_products",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetAllCategoriesModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
        "/v1/category/{category_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "API for getting a category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category"
                ],
                "summary": "Get Category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Shipper",
                        "name": "Shipper",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "category_id",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetCategoryModel"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ChildCategoryModel": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "description_v2": {
                    "$ref": "#/definitions/models.Language"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string",
                    "example": "0"
                },
                "parent_id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetProductModel"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/models.Language"
                }
            }
        },
...
        "models.GetAllCategoriesModel": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetCategoryModel"
                    }
                },
                "count": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "models.GetCategoryModel": {
            "type": "object",
            "properties": {
                "child_categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ChildCategoryModel"
                    }
                },
                "description": {
                    "type": "string"
                },
                "description_v2": {
                    "$ref": "#/definitions/models.Language"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string",
                    "example": "0"
                },
                "parent_id": {
                    "type": "string"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.GetProductModel"
                    }
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/models.Language"
                }
            }
        },
...
        "models.GetProductModel": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "description_v2": {
                    "$ref": "#/definitions/models.Language"
                },
                "id": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "order_no": {
                    "type": "string",
                    "example": "0"
                },
                "price": {
                    "type": "string",
                    "example": "0"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "$ref": "#/definitions/models.Language"
                }
            }
        },
...
        "models.Language": {
            "type": "object",
            "properties": {
                "en": {
                    "type": "string"
                },
                "ru": {
                    "type": "string"
                },
                "uz": {
                    "type": "string"
                }
            }
        },
...
        "models.ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}